add caching to geom drugs
