+++ atomic_datasets/utils/__init__.py
@@ -7,3 +7,4 @@ from .rdkit import (
 )
 from .download import clone_url, download_url, extract_zip, extract_gz, extract_tar
 from .decorators import after_preprocess
+from .cache import cache_to_file
